src/requirements.txt
PyYAML
msgpack-python
pymongo
pyzmq
redis
motor
python-box
requests
lxml
pystache